+++ index.js
@@ -1,9 +1,7 @@
 'use strict';
 
-let fetch = require('node-fetch');
 let BotWrapper = require('./lib/BotWrapper.js');
 let ftl = require('./lib/ftl.js');
-let stripContent = require('./lib/utils').stripContent;
 
 let pingCount = 0;
 
@@ -36,43 +34,7 @@ bot.registerCommand('ping', () => {
 });
 
 require('./commands/codeOfConduct').onJoin({ bot });
-
-bot.registerCommand('news', async (msg) => {
-  const newsChannel = process.env.NEWS_CHANNEL;
-  const contentForNewsChannel = stripContent(msg.content);
-
-  let userName = 'unknown';
-  let messageChannelName = 'unknown';
-  if (msg.channel.type == 0) {
-    userName = msg.member.nick ? msg.member.nick : msg.author.username;
-    messageChannelName = msg.channel.name;
-  } else if (msg.channel.type == 1) {
-    userName = msg.author.username;
-    messageChannelName = ftl('news-dm-description');
-  }
-
-  let content = ftl('news-post-message', {
-    userName,
-    messageChannelName,
-    contentForNewsChannel,
-  });
-
-  let files = await Promise.all(
-    msg.attachments.map(async (attachment) => {
-      try {
-        let res = await fetch(attachment.url);
-        let buffer = await res.buffer();
-        return { file: buffer, name: attachment.filename };
-      } catch (err) {
-        console.warning(`Couldn't fetch attachment from message ${msg.id}: ${attachment.url}`);
-        return null;
-      }
-    }),
-  );
-  files = files.filter((a) => a !== null);
-  bot.createMessage(newsChannel, content, files);
-});
-
+require('./commands/newsFeed').register({ bot });
 require('./commands/diceRoller').register({ bot });
 require('./commands/codeOfConduct').register({ bot });
