+++ lib/utils.js
@@ -1,3 +1,4 @@
+let ftl = require('../lib/ftl');
 // General functions that are command agnostic.
 module.exports.stripContent = (messageContent) => {
   const stringParts = messageContent.split(' ');
@@ -5,4 +6,22 @@ module.exports.stripContent = (messageContent) => {
 
   const userPost = stringParts.join(' ');
   return userPost;
+};
+
+module.exports.memberName = (msg) => {
+  if (msg.channel.type === 0) {
+    return msg.member.nick ? msg.member.nick : msg.author.username; //Prefers member nickname when available
+  } else if (msg.channel.type === 1) {
+    return msg.author.username; //Nickname not available from DMs
+  }
+  return 'unknown';
+};
+
+module.exports.incomingChannel = (msg) => {
+  if (msg.channel.type === 0) {
+    return msg.channel.name;
+  } else if (msg.channel.type === 1) {
+    return ftl('news-dm-description'); //There is no built in name for the DM channel
+  }
+  return 'unknown';
 };
